Add functions for handling data response in chat module
