+++ utils/app/chat.ts
@@ -8,6 +8,19 @@ import {
   DEFAULT_IMAGE_GENERATION_QUALITY,
   DEFAULT_IMAGE_GENERATION_STYLE,
 } from './const';
+import {
+  saveConversation,
+  saveConversations,
+  updateConversationLastUpdatedAtTimeStamp,
+} from './conversation';
+import {
+  removeRedundantTempHtmlString,
+  removeTempHtmlString,
+} from './htmlStringHandler';
+import { reorderItem } from './rank';
+import { removeSecondLastLine } from './ui';
+
+import dayjs from 'dayjs';
 
 function updateConversation(
   deleteCount: number,
@@ -132,12 +145,160 @@ function handleNoDataResponse(homeDispatch: Function) {
   homeDispatch({ field: 'messageIsStreaming', value: false });
 }
 
+async function handleDataResponse(
+  data: ReadableStream<Uint8Array>,
+  updatedConversation: Conversation,
+  plugin: Plugin | null,
+  message: Message,
+  controller: AbortController,
+  selectedConversation: Conversation,
+  conversations: Conversation[],
+  stopConversationRef: React.MutableRefObject<boolean>,
+  homeDispatch: Function,
+) {
+  if (updatedConversation.messages.length === 1) {
+    const { content } = message;
+    const customName =
+      content.length > 30 ? content.substring(0, 30) + '...' : content;
+    updatedConversation = {
+      ...updatedConversation,
+      name: customName,
+    };
+  }
+  homeDispatch({ field: 'loading', value: false });
+  const reader = data.getReader();
+  const decoder = new TextDecoder();
+  let done = false;
+  let isFirst = true;
+  let text = '';
+  let largeContextResponse = false;
+  let showHintForLargeContextResponse = false;
+
+  while (!done) {
+    if (stopConversationRef.current === true) {
+      controller.abort();
+      done = true;
+      break;
+    }
+    const { value, done: doneReading } = await reader.read();
+    done = doneReading;
+    const chunkValue = decoder.decode(value);
+    text += chunkValue;
+
+    if (text.includes('[DONE]')) {
+      text = text.replace('[DONE]', '');
+      done = true;
+    }
+
+    if (text.includes('[16K]')) {
+      text = text.replace('[16K]', '');
+      largeContextResponse = true;
+    }
+
+    if (text.includes('[16K-Optional]')) {
+      text = text.replace('[16K-Optional]', '');
+      showHintForLargeContextResponse = true;
+    }
+
+    if (text.includes('[REMOVE_TEMP_HTML]')) {
+      text = removeTempHtmlString(text);
+    }
+
+    if (text.includes('[REMOVE_LAST_LINE]')) {
+      text = text.replace('[REMOVE_LAST_LINE]', '');
+      text = removeSecondLastLine(text);
+    }
+
+    // We can use this command to trigger the initial stream of Edge function response
+    // so we have more than 25 seconds on Vercel Edge network to wait for response
+    if (text.includes('[PLACEHOLDER]')) {
+      text = text.replace('[PLACEHOLDER]', '');
+    }
+
+    if (isFirst) {
+      isFirst = false;
+      const updatedMessages: Message[] = [
+        ...updatedConversation.messages,
+        {
+          role: 'assistant',
+          content: removeRedundantTempHtmlString(text),
+          largeContextResponse,
+          showHintForLargeContextResponse,
+          pluginId: plugin?.id || null,
+        },
+      ];
+      updatedConversation = {
+        ...updatedConversation,
+        messages: updatedMessages,
+        lastUpdateAtUTC: dayjs().valueOf(),
+      };
+      homeDispatch({
+        field: 'selectedConversation',
+        value: updatedConversation,
+      });
+    } else {
+      const updatedMessages: Message[] = updatedConversation.messages.map(
+        (message, index) => {
+          if (index === updatedConversation.messages.length - 1) {
+            return {
+              ...message,
+              content: removeRedundantTempHtmlString(text),
+              largeContextResponse,
+              showHintForLargeContextResponse,
+            };
+          }
+          return message;
+        },
+      );
+      updatedConversation = {
+        ...updatedConversation,
+        messages: updatedMessages,
+        lastUpdateAtUTC: dayjs().valueOf(),
+      };
+      homeDispatch({
+        field: 'selectedConversation',
+        value: updatedConversation,
+      });
+    }
+  }
+
+  saveConversation(updatedConversation);
+  let updatedConversations: Conversation[] = conversations.map(
+    (conversation) => {
+      if (conversation.id === selectedConversation.id) {
+        return updatedConversation;
+      }
+      return conversation;
+    },
+  );
+
+  // If the conversation is new, add it to the list of conversations
+  if (
+    !updatedConversations.find(
+      (conversation) => conversation.id === updatedConversation.id,
+    )
+  ) {
+    updatedConversations.push(updatedConversation);
+    updatedConversations = reorderItem(
+      updatedConversations,
+      updatedConversation.id,
+      0,
+    );
+  }
+
+  homeDispatch({ field: 'conversations', value: updatedConversations });
+  saveConversations(updatedConversations);
+  homeDispatch({ field: 'messageIsStreaming', value: false });
+
+  updateConversationLastUpdatedAtTimeStamp();
+}
 const chat = {
   updateConversation,
   createChatBody,
   sendRequest,
   handleErrorResponse,
   handleNoDataResponse,
+  handleDataResponse,
 };
 
 export default chat;
